+++ CHANGELOG.md
@@ -7390,6 +7390,8 @@ This reverts commit fd44da6e4d5adea0a7153b0d6f7f0b7334ab0c61.
 
 ### New
 
+- editors/neovim: add custom spellfile
+
 - treewide: add "vim" to luarc globals
 
 - flake: add exiftools hook; rename lib.nix to utils.nix